pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pygmdl"
version = "1.0.0"
description = "Download satellite images from Google Maps."
authors = [{name = "iwatkot", email = "iwatkot@gmail.com"}]
license = {text = "MIT License"}
readme = "README.md"
keywords = ["googlemaps", "satellite", "images", "download"]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow",
    "pyproj",
    "geopy",
]

[project.urls]
Homepage = "https://github.com/iwatkot/pygmdl"
Repository = "https://github.com/iwatkot/pygmdl"

[tool.setuptools.packages.find]
where = ["."]
include = ["pygmdl*"]
exclude = ["dev*"]
